pyproject.toml
[project]
name = "job-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai[all]>=0.8.0",
    "google-genai>=1.61.0",
    "litellm>=1.81.5",
    "mlflow>=2.14.0",
    "notebook>=7.5.3",
    "pandas<3.0.0",
    "pip>=25.3",
    "playwright>=1.57.0",
    "protobuf<5.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "structlog>=25.5.0",
    "tenacity>=9.1.2",
    "unidecode>=1.4.0",
]
